Prevent use of the old bad default NTP server.
This prevents the use of the old bad default 50.97.210.169 NTP server by scrubbing that value at config load time.

Why? https://opensprinkler.com/forums/topic/wrong-time-on-the-unit/ and related threads and support tickets about that problem that don't need to happen. With this code in place in a future release people upgrading from older versions won't experience the problem.

I do not have a system setup to build and test this firmware.  Consider this a proposal for one thing that could save you support issues in the future.  If you want to accept it, you'll need to do the rest.

As this just runs the fixup at config load time, someone could still set it and save it and it'd be active in the running controller until reboot.  Not ideal?  Doing config fixups as their own function that is called at the end of iopts_load() and the beginning of iopts_save() might be nicer and avoids such inconsistency.

This value scrubbing could also be done in the App when saving or restoring a config.  _(I cannot personally make PRs to the app due to its license, AGPLed code is off-limits for me - so that'll be up to someone else)._