Make firmware PlatformIO compilable
Adds the needed `platformio.ini` plus a refactor of the `server.h` and
`server.cpp` files because they confict with the Arduino-ESP8266 core
files of `Server.h` and friends. This allows it to properly build on
Windows which has a case-insenstive filenames and the wrong header file
would be picked up, leading to compilation errors.

Folder structure is otherwise unchanged.

`platformio.ini` currently only builds the ESP8266 version of the
fimware, but can also be expanded to build the AVR version of it.